requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
click==8.0.3; python_version >= '3.6'
colorama==0.4.4; platform_system == 'Windows'
dnspython==2.1.0
flask-pymongo==2.3.0
flask==2.0.2
gunicorn==20.1.0
itsdangerous==2.0.1; python_version >= '3.6'
jinja2==3.0.3; python_version >= '3.6'
markupsafe==2.0.1; python_version >= '3.6'
pymongo[srv]==4.0.1
python-dotenv==0.19.2
werkzeug==2.0.2; python_version >= '3.6'
